+++ pkg/_macros/lib/src/api/code.dart
@@ -7,8 +7,12 @@ part of '../api.dart';
 /// The base class representing an arbitrary chunk of Dart code, which may or
 /// may not be syntactically or semantically valid yet.
 sealed class Code {
-  /// All the chunks of [Code], raw [String]s, or [Identifier]s that
-  /// comprise this [Code] object.
+  /// All the chunks of [Code], raw [String]s, [Identifier]s, or
+  /// [OmittedTypeAnnotation]s that comprise this [Code] object.
+  ///
+  /// Note that [OmittedTypeAnnotation] objects can only be provided through
+  /// the [OmittedTypeAnnotationCode] wrapper instance, but will appear in
+  /// the [parts] of those, so they must be handled whenever iterating [parts].
   final List<Object> parts;
 
   /// Can be used to more efficiently detect the kind of code, avoiding is