+++ tests/language/wildcard_variables/non_binding/top_level_variable_test.dart
@@ -15,16 +15,21 @@ void main() {
   Clas().member();
 
   int _ = _;
-  int _ = 2;
-  _ = 3;
-  Expect.equals(3, _);
+  int _ = 3;
+  var (_, _) = (3, '4');
+  Expect.equals(2, _);
+
+  _ = 4;
+  Expect.equals(4, _);
 }
 
 class Clas<_> {
   void member<_>() {
     int _ = _;
-    int _ = 2;
+    int _ = 3;
+    var (_, _) = (3, '4');
     Expect.equals(2, _);
+
     _ = 4;
     Expect.equals(4, _);